+++ lib/widget/battle/detail/notice.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:poorlex/controller/image_picker_controller.dart';
+import 'package:poorlex/controller/image_picker.dart';
 
 import 'package:poorlex/libs/theme.dart';
